Add more tests for ToastContext
